refactor(media): relocate music duration contract
